Attempt to fix crash in push channel
The push channel connection would sometimes assert in dealloc because it
wasn't closed. After inspecting the threads involved I think this is a
race condition created by the facts that the push channel methods are
executed on both the sync thread and transport session thread.

This PR re-factor the push channel to execute only on the transport
session thread.